one/api/requirements.txt
aiofiles==0.6.0
alembic==1.5.8
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.22.0
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
databases==0.4.2
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.0.0
h11==0.12.0
httpcore==0.12.3
httptools==0.1.1
httpx==0.17.1
idna==2.10
iniconfig==1.1.1
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
orjson==3.5.1
packaging==20.9
paramiko==2.7.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
PyJWT==2.0.1
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.3
pytest-asyncio==0.14.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
Rx==1.6.1
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.3
SQLAlchemy-Utils==0.37.0
sshtunnel==0.4.0
starlette==0.13.6
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
